add matrix type

+++ laboratory-2/src/maze_placeholder.py
@@ -23,5 +23,6 @@ MAZE_MAP_CLASSIC = np.asarray(
         [W, W, W, W, W, W, W, W, W, W, W, W, W, W, W, W, W, W],
         [W, W, W, W, W, W, W, W, W, W, W, W, W, W, W, W, W, W],
         [W, W, W, W, W, W, W, W, W, W, W, W, W, W, W, W, W, W],
-    ]
+    ],
+    dtype=np.uint8,
 )
